Split test files and prepare for Rails 6.1 support
Rails 6.1 introduced legacy_connection_handling which lets you
opt-in/out of the new multiple database support functionality.

A lot of our ARHP tests are written for the legacy connection handling
and so this commit splits those into their own file.

Other tests we want to run irregardless of legacy_connection_handling
status but we need to load the new database format in
legacy_connection_handling=false scenarios.
