+++ test/helper.rb
@@ -9,6 +9,7 @@ require 'mocha/minitest'
 require 'phenix'
 
 RAILS_ENV = 'test'
+ENV['LEGACY_CONNECTION_HANDLING'] = 'true' if ENV['LEGACY_CONNECTION_HANDLING'].nil?
 
 if ActiveRecord.version >= Gem::Version.new('6.1')
   ActiveRecord::Base.legacy_connection_handling = (ENV['LEGACY_CONNECTION_HANDLING'] == 'true')
