+++ test/test_arhp.rb
@@ -5,7 +5,11 @@ require_relative 'helper'
 class ActiveRecordHostPoolTest < Minitest::Test
   include ARHPTestSetup
   def setup
-    Phenix.rise!
+    if ActiveRecord.version >= Gem::Version.new('6.1') && !ActiveRecord::Base.legacy_connection_handling
+      Phenix.rise! config_path: 'test/three_tier_database.yml'
+    else
+      Phenix.rise!
+    end
     arhp_create_models
   end
 
@@ -43,10 +47,6 @@ class ActiveRecordHostPoolTest < Minitest::Test
     refute_equal(Pool2DbE.connection.raw_connection, Pool3DbE.connection.raw_connection)
   end
 
-  def test_models_with_different_replica_status_should_not_share_a_connection
-    refute_equal(Pool1DbA.connection.raw_connection, Pool1DbAReplica.connection.raw_connection)
-  end
-
   def test_should_select_on_correct_database
     Pool1DbA.connection.send(:select_all, 'select 1')
     assert_equal 'arhp_test_db_a', current_database(Pool1DbA)
@@ -67,38 +67,6 @@ class ActiveRecordHostPoolTest < Minitest::Test
 
     Pool3DbE.connection.send(:insert, "insert into tests values(NULL, 'foo')")
     assert_equal 'arhp_test_db_e', current_database(Pool3DbE)
-  end
-
-  def test_models_with_matching_hosts_and_non_matching_databases_should_share_a_connection
-    simulate_rails_app_active_record_railties
-    assert_equal(Pool1DbA.connection.raw_connection, Pool1DbC.connection.raw_connection)
-  end
-
-  if ActiveRecord.version >= Gem::Version.new('6.0')
-    def test_models_with_matching_hosts_and_non_matching_databases_issue_exists_without_arhp_patch
-      simulate_rails_app_active_record_railties
-
-      # Remove patch that fixes an issue in Rails 6+ to ensure it still
-      # exists. If this begins to fail then it may mean that Rails has fixed
-      # the issue so that it no longer occurs.
-      without_module_patch(ActiveRecordHostPool::ClearQueryCachePatch, :clear_query_caches_for_current_thread) do
-        exception = assert_raises(ActiveRecord::StatementInvalid) do
-          ActiveRecord::Base.cache { Pool1DbC.create! }
-        end
-
-        assert_equal("Mysql2::Error: Table 'arhp_test_db_b.pool1_db_cs' doesn't exist", exception.message)
-      end
-    end
-
-    def test_models_with_matching_hosts_and_non_matching_databases_do_not_mix_up_underlying_database
-      simulate_rails_app_active_record_railties
-
-      # ActiveRecord 6.0 introduced a change that surfaced a problematic code
-      # path in active_record_host_pool when clearing caches across connection
-      # handlers which can cause the database to change.
-      # See ActiveRecordHostPool::ClearQueryCachePatch
-      ActiveRecord::Base.cache { Pool1DbC.create! }
-    end
   end
 
   def test_connection_returns_a_proxy
@@ -163,7 +131,7 @@ class ActiveRecordHostPoolTest < Minitest::Test
     conn.expects(:execute_without_switching)
     conn.expects(:_switch_connection).never
     assert conn._host_pool_current_database
-    conn.create_database(:some_args)
+    conn.create_database(:some_args, charset: 'utf8mb4')
   end
 
   def test_no_switch_when_dropping_db
@@ -214,19 +182,5 @@ class ActiveRecordHostPoolTest < Minitest::Test
     conn = pool.connection
     pool.expects(:checkin).with(conn)
     pool.release_connection
-  end
-
-  private
-
-  def simulate_rails_app_active_record_railties
-    if ActiveRecord.version >= Gem::Version.new('6.0')
-      # Necessary for testing ActiveRecord 6.0 which uses the connection
-      # handlers when clearing query caches across all handlers when
-      # an operation that dirties the cache is involved (e.g. create/insert,
-      # update, delete/destroy, truncate, etc.)
-      ActiveRecord::Base.connection_handlers = {
-        ActiveRecord::Base.writing_role => ActiveRecord::Base.default_connection_handler
-      }
-    end
   end
 end
